Adding Policies, fixing bugs.

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -10,6 +10,8 @@ class CategoryController extends Controller
 {
     public function index()
     {
+        $this->authorize('view', Category::class);
+
         $categories = Category::all();   
         $categories = $categories->map(function ($category) {
             $category->title = ucwords($category->title);
@@ -36,17 +38,23 @@ class CategoryController extends Controller
 
     public function edit(Category $category)
     {
+        $this->authorize('update', $category);
+
         return view('admin.categories', compact('category'));
     }
 
     public function update(Category $category, CategoryRequest $request)
     {
+        $this->authorize('update', $category);
+
         $category->update($request->all());
         return redirect('admin/categories')->with('success', __('success.update_category') . ' ' . $request['title']); 
     }
 
     public function destroy(Category $category)
     {
+        $this->authorize('delete', $category);
+
         $category->delete();
         return redirect('admin/categories')->with('success', __('success.delete_category'));
     }
